Implemented Vivado version switching using new VersionRange class from pyTooling.
