+++ tests/requirements.txt
@@ -8,6 +8,6 @@ pytest ~= 8.4
 pytest-cov ~= 6.2
 
 # Static Type Checking
-mypy ~= 1.16
+mypy ~= 1.17
 typing_extensions ~= 4.14
 lxml ~= 6.0
